implement makeListEvents()
This is just the first implementation.
I want to start implementing some routes,
but didn't want the first route to be a POST one.
Also figured out how to test a function
that returns some data.
I want to add some 'mappers' to the createEvent() and listEvent() results.
Not sure where yet.